pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager made with Django"
authors = ["Nedashkovsky <Nedashkovsky@ymail.com>"]
readme = "README.md"
packages = [
    {include = "task_manager"}
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
gunicorn = "^20.1.0"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.3"
django-bootstrap4 = "^22.2"
python-gettext = "^4.1"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"
rollbar = "^0.16.3"
dj-database-url = "0.5.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
